内部のモジュールを ES Modules をデフォルトになるよう書き換える
