+++ src/+react.cjs
