+++ src/flat/configs/react.js
@@ -2,8 +2,8 @@ import { fixupConfigRules } from '@eslint/compat';
 import reactPlugin from 'eslint-plugin-react';
 import globals from 'globals';
 
-import legacyReact from '../../+react.js';
-import { compat, jsPattern, tsPattern } from '../util.mjs';
+import legacyReact from '../../+react.cjs';
+import { compat, jsPattern, tsPattern } from '../util.js';
 
 export const reactConfigs = /** @satisfies {import('eslint').Linter.Config[]} */ ([
   reactPlugin.configs.flat.recommended,
