+++ src/flat/configs/typescript.js
@@ -1,7 +1,7 @@
 import tseslint from 'typescript-eslint';
 
-import legacyTypeScriptConfig from '../../+typescript.js';
-import { tsPattern } from '../util.mjs';
+import legacyTypeScriptConfig from '../../+typescript.cjs';
+import { tsPattern } from '../util.js';
 
 // https://github.com/typescript-eslint/typescript-eslint/issues/9724 回避のためにキャストする
 const recommendedTypeChecked = /** @type {import('eslint').Linter.Config[]} */ (
