+++ src/flat/index.js
@@ -1,0 +1,15 @@
+import { baseConfigs } from './configs/base.js';
+import { nodeConfigs } from './configs/node.js';
+import { prettierConfig } from './configs/prettier.js';
+import { reactConfigs } from './configs/react.js';
+import { typescriptConfigs } from './configs/typescript.js';
+
+const mizdra = {
+  baseConfigs,
+  typescriptConfigs,
+  reactConfigs,
+  nodeConfigs,
+  prettierConfig,
+};
+
+export default mizdra;
