+++ test/eslint.config.js
@@ -1,11 +1,13 @@
 // @ts-check
-import mizdra from './src/flat/index.mjs';
+
+import mizdra from '../src/flat/index.js';
 
 /** @type {import('eslint').Linter.Config[]} */
 export default [
-  { ignores: ['test'] },
+  { ignores: ['.eslintrc.cjs', 'eslint.config.js'] },
   ...mizdra.baseConfigs,
   ...mizdra.typescriptConfigs,
   ...mizdra.nodeConfigs,
+  ...mizdra.reactConfigs,
   mizdra.prettierConfig,
 ];
